+++ src/solutions/problem18.rs
@@ -10,7 +10,7 @@ pub fn solution() -> SolutionResult {
     let mut lines = contents[].lines();
     let mut rows = Vec::new();
     for line in lines {
-        let row = line.words().filter_map(|s| s.parse()).collect::<Vec<i32>>();
+        let row = line.words().filter_map(|s| s.parse()).collect::<Vec<i64>>();
         rows.push(row);
     }
     for i in range(0, rows.len() - 1).rev() {
