+++ src/solutions/problem19.rs
@@ -1,11 +1,11 @@
 use super::SolutionResult;
 
 
-fn is_leap_year(year: i32) -> bool {
+fn is_leap_year(year: i64) -> bool {
     (year % 4 == 0) && (year % 100 != 0 || year % 400 == 0)
 }
 
-fn days_in_february(year: i32) -> i32 {
+fn days_in_february(year: i64) -> i64 {
     if is_leap_year(year) {
         29
     } else {
@@ -13,17 +13,17 @@ fn days_in_february(year: i32) -> i32 {
     }
 }
 
-fn year_days(year: i32) -> Vec<i32> {
+fn year_days(year: i64) -> Vec<i64> {
     vec!(31, days_in_february(year), 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
 }
 
 pub fn solution() -> SolutionResult {
-    let days = range(1901i32, 2001).flat_map(|year| year_days(year).into_iter());
-    let num_sundays = days.scan(0, |acc: &mut i32, x: i32| {
+    let days = range(1901i64, 2001).flat_map(|year| year_days(year).into_iter());
+    let num_sundays = days.scan(0, |acc: &mut i64, x: i64| {
         *acc = *acc + x;
         Some(*acc)
     }).filter(|&x| x % 7 == 0).count() + 1;
-    Ok(num_sundays as i32)
+    Ok(num_sundays as i64)
 }
 
 #[cfg(test)]
