+++ src/solutions/problem2.rs
@@ -4,7 +4,7 @@ use super::SolutionResult;
 use super::common::Fib;
 
 pub fn solution() -> SolutionResult {
-    let fib: Fib<i32> = Fib::new();
+    let fib: Fib<i64> = Fib::new();
     Ok(fib.take_while(|&i| i <= 4000000).filter(|&i| i % 2 == 0).sum())
 }
