+++ src/solutions/problem21.rs
@@ -4,16 +4,16 @@ use std::iter::AdditiveIterator;
 use super::SolutionResult;
 use super::common::divisors;
 
-fn sum_divisors(n: i32) -> i32 {
+fn sum_divisors(n: i64) -> i64 {
     divisors(n).into_iter().sum()
 }
 
 pub fn solution() -> SolutionResult {
     let mut divisors_map = HashMap::new();
-    for i in range(2i32, 10000) {
+    for i in range(2i64, 10000) {
         divisors_map.insert(i, sum_divisors(i));
     }
-    let mut sum = 0i32;
+    let mut sum = 0i64;
     for (key, value) in divisors_map.iter() {
         if *key != *value && divisors_map.get(value) == Some(key) {
             sum += *value;
