+++ src/solutions/problem22.rs
@@ -17,7 +17,7 @@ pub fn solution() -> SolutionResult {
     let contents = File::open(&path).read_to_string().ok().expect("Cannot read file");
     let mut names: Vec<&str> = contents[].split(',').collect();
     names.sort();
-    Ok(names.into_iter().enumerate().map(|(i, name)| (i as u32 + 1) * name_value(name)).sum() as i32)
+    Ok(names.into_iter().enumerate().map(|(i, name)| (i as u32 + 1) * name_value(name)).sum() as i64)
 }
 
 #[cfg(test)]
