+++ src/solutions/problem34.rs
@@ -4,12 +4,12 @@ use std::iter::MultiplicativeIterator;
 use super::common::digits;
 use super::SolutionResult;
 
-fn factorial(n: i32) -> i32 {
+fn factorial(n: i64) -> i64 {
     return range(2, n + 1).product()
 }
 
 pub fn solution() -> SolutionResult {
-    Ok(range(3i32, 2540160).filter(|&x| x == digits(x).into_iter().map(factorial).sum()).sum())
+    Ok(range(3i64, 2540160).filter(|&x| x == digits(x).into_iter().map(factorial).sum()).sum())
 }
 
 #[cfg(test)]
