+++ src/solutions/problem7.rs
@@ -6,7 +6,7 @@ pub fn solution() -> SolutionResult {
     let (_, upper_bound) = slow_primes::estimate_nth_prime(10001);
     let primes = Primes::sieve(upper_bound as usize);
     match primes.primes().nth(10001 - 1) {
-        Some(p) => Ok(p as i32),
+        Some(p) => Ok(p as i64),
         None => Err(SolutionError::MatchFailed),
     }
 }
